+++ NEWS.md
@@ -201,6 +201,8 @@ Deprecated or removed
 
   * The method `String(io::IOBuffer)` is deprecated to `String(take!(copy(io)))` ([#21438]).
 
+  * The function `readstring` is deprecated in favor of `read(io, String)` ([#22793])
+
   * The function `showall` is deprecated. Showing entire values is the default, unless an
     `IOContext` specifying `:limit=>true` is in use ([#22847]).